test multipart upload cycle

+++ FileService/src/FileService.Contracts/Responses/StartMultipartUploadResponse.cs
@@ -5,7 +5,7 @@ public sealed record StartMultipartUploadResponse
     public Guid MediaAssetId { get; init; } 
     public required string UploadId { get; init; }
     public required List<ChunkUploadUrlDto> ChunkUrls { get; init; }
-    public long ChunkSize { get; init; }
+    public int ChunkSize { get; init; }
 }
 
 public sealed record ChunkUploadUrlDto
